backend/requirements.txt
fastapi
uvicorn[standard]
openai>=1.30.0
matplotlib
pydantic>=2.0
python-dotenv
sse-starlette
